hallucination.py
import torch
from transformers import GPT2LMHeadModel, GPT2Tokenizer
import faiss
import numpy as np
import logging
from sklearn.metrics.pairwise import cosine_similarity

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Load GPT2 model and tokenizer
tokenizer = GPT2Tokenizer.from_pretrained("gpt2")
model = GPT2LMHeadModel.from_pretrained("gpt2")

# Dummy knowledge base for fact-checking (example, replace with real KB)
knowledge_base = [
    "Python is a programming language.",
    "TensorFlow is a popular deep learning framework.",
    "OpenAI developed GPT models."
]

# FAISS setup for retrieving relevant documents from the knowledge base
def setup_faiss(knowledge_base):
    # Create a FAISS index for fast similarity search
    dim = 512  # Embedding dimension for retrieval
    faiss_index = faiss.IndexFlatL2(dim)  # L2 distance for similarity search
    
    # Dummy embedding for each knowledge base entry
    embeddings = np.random.rand(len(knowledge_base), dim).astype("float32")  # Replace with real embeddings
    faiss_index.add(embeddings)
    
    return faiss_index, knowledge_base

# Fact-checking function using FAISS and cosine similarity
def fact_check(query, faiss_index, knowledge_base):
    # Tokenize and encode the query
    inputs = tokenizer(query, return_tensors="pt")
    query_embedding = model.transformer.wte(inputs["input_ids"]).mean(dim=1).detach().numpy().astype("float32")
    
    # Perform FAISS similarity search
    _, I = faiss_index.search(query_embedding, k=3)  # Top 3 documents
    retrieved_documents = [knowledge_base[i] for i in I[0]]
    
    # Simulate fact-checking by checking if any retrieved document matches the query's intent
    for doc in retrieved_documents:
        similarity = cosine_similarity(query_embedding, np.random.rand(1, 512))  # Replace with real comparison logic
        if similarity > 0.8:  # Threshold for similarity
            return doc
    return "Unable to validate the information."

# Model generation with hallucination detection
def generate_with_fact_checking(query, faiss_index, knowledge_base):
    # Fact-check the query first
    retrieved_fact = fact_check(query, faiss_index, knowledge_base)
    logger.info(f"Fact-Checked Information: {retrieved_fact}")
    
    # If fact-checking passes, generate response
    inputs = tokenizer(query, return_tensors="pt")
    outputs = model.generate(inputs["input_ids"], max_length=50, num_return_sequences=1)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    
    # Simple hallucination check (for demonstration purposes)
    if "error" in generated_text or "incorrect" in generated_text:
        return "Sorry, I couldn't find a valid answer."
    return generated_text

# Example of setting up FAISS and knowledge base
faiss_index, knowledge_base = setup_faiss(knowledge_base)

# Sample query
query = "What is Python?"

# Generate response with fact-checking and hallucination mitigation
response = generate_with_fact_checking(query, faiss_index, knowledge_base)
print(response)
